+++ Runtime/AnimationRig/Constraints/ChainIKConstraint.cs
@@ -59,7 +59,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Chain IK Constraint")]
-    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/ChainIKConstraint.html")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@0.2?preview=1&subfolder=/manual/constraints/ChainIKConstraint.html")]
     public class ChainIKConstraint : RigConstraint<
         ChainIKConstraintJob,
         ChainIKConstraintData,