+++ Runtime/AnimationRig/Constraints/MultiAimConstraint.cs
@@ -98,7 +98,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Multi-Aim Constraint")]
-    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/MultiAimConstraint.html")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@0.2?preview=1&subfolder=/manual/constraints/MultiAimConstraint.html")]
     public class MultiAimConstraint : RigConstraint<
         MultiAimConstraintJob,
         MultiAimConstraintData,