+++ Runtime/AnimationRig/Constraints/MultiReferentialConstraint.cs
@@ -58,7 +58,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Multi-Referential Constraint")]
-    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/MultiReferentialConstraint.html")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@0.2?preview=1&subfolder=/manual/constraints/MultiReferentialConstraint.html")]
     public class MultiReferentialConstraint : RigConstraint<
         MultiReferentialConstraintJob,
         MultiReferentialConstraintData,